Extract business logic in ShareViewModel to domain layer
